refactor: Change the path of 'Size' related files to 'Geometry'
