+++ Includes/Core/Grid/Grid2.hpp
@@ -12,7 +12,7 @@
 #define CUBBYFLOW_GRID2_HPP
 
 #include <Core/Geometry/BoundingBox2.hpp>
-#include <Core/Size/Size2.hpp>
+#include <Core/Geometry/Size2.hpp>
 #include <Core/Utils/Serialization.hpp>
 #include <Core/Vector/Vector2.hpp>
