+++ Includes/Core/Searcher/PointHashGridSearcher2.hpp
@@ -11,8 +11,8 @@
 #ifndef CUBBYFLOW_POINT_HASH_GRID_SEARCHER2_HPP
 #define CUBBYFLOW_POINT_HASH_GRID_SEARCHER2_HPP
 
+#include <Core/Geometry/Size2.hpp>
 #include <Core/Searcher/PointNeighborSearcher2.hpp>
-#include <Core/Size/Size2.hpp>
 
 #include <vector>
