+++ Includes/Core/Solver/Grid/GridBoundaryConditionSolver3.hpp
@@ -13,8 +13,8 @@
 
 #include <Core/Field/ScalarField3.hpp>
 #include <Core/Geometry/Collider3.hpp>
+#include <Core/Geometry/Size3.hpp>
 #include <Core/Grid/FaceCenteredGrid3.hpp>
-#include <Core/Size/Size3.hpp>
 
 namespace CubbyFlow
 {
